+++ cyxbs-pages/course/src/mobileMain/kotlin/com/cyxbs/pages/course/frame/MobileHomeCourseFrame.kt
@@ -18,12 +18,16 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.cyxbs.components.config.compose.theme.LocalAppColors
 import com.cyxbs.components.view.ui.BottomSheetState
 import com.cyxbs.pages.course.api.IMobileHomeCourseFrame
+import com.cyxbs.pages.course.frame.bottomsheet.MobileHomeBottomSheet
 import com.cyxbs.pages.course.frame.decoration.AffairDecorationViewModel
 import com.cyxbs.pages.course.frame.decoration.LinkLessonDecorationViewModel
 import com.cyxbs.pages.course.frame.decoration.SelfLessonDecorationViewModel
-import com.cyxbs.pages.course.frame.bottomsheet.MobileHomeBottomSheet
 import com.cyxbs.pages.course.frame.header.MobileHomeCourseHeader
+import com.cyxbs.pages.course.frame.item.CourseAffairItem
+import com.cyxbs.pages.course.frame.item.LinkLessonItem
+import com.cyxbs.pages.course.frame.item.SelfLessonItem
 import com.cyxbs.pages.course.view.decoration.CoursePageDecoration
+import com.cyxbs.pages.course.view.item.CourseItemHierarchy
 import com.cyxbs.pages.course.view.item.CourseItemViewModel
 import com.cyxbs.pages.course.view.item.affair.CreateAffairDecorationViewModel
 import com.g985892345.provider.api.annotation.ImplProvider
@@ -102,23 +106,19 @@ private fun MobileHomeCourseFrameContent(
 
 @Composable
 private fun getDecoration(): ImmutableList<CoursePageDecoration> {
-  val courseItemViewModel = viewModel { CourseItemViewModel() }
-  val selfLessonDecoration = viewModel {
-    SelfLessonDecorationViewModel(
-      hierarchy = courseItemViewModel.createOverlay()
-    )
-  }
-  val linkLessonDecoration = viewModel {
-    LinkLessonDecorationViewModel(
-      hierarchy = courseItemViewModel.createOverlay()
-    )
-  }
-  val affairDecoration = viewModel {
-    AffairDecorationViewModel(
-      hierarchy = courseItemViewModel.createOverlay()
-    )
-  }
+  val selfLessonHierarchy = remember { CourseItemHierarchy<SelfLessonItem>() }
+  val selfLessonDecoration = viewModel { SelfLessonDecorationViewModel(selfLessonHierarchy) }
+
+  val linkLessonHierarchy = remember { CourseItemHierarchy<LinkLessonItem>() }
+  val linkLessonDecoration = viewModel { LinkLessonDecorationViewModel(linkLessonHierarchy) }
+
+  val affairHierarchy = remember { CourseItemHierarchy<CourseAffairItem>() }
+  val affairDecoration = viewModel { AffairDecorationViewModel(affairHierarchy) }
+
   val createAffairDecoration = viewModel { CreateAffairDecorationViewModel() }
+
+
+  val courseItemViewModel = viewModel { CourseItemViewModel(selfLessonHierarchy, linkLessonHierarchy, affairHierarchy) }
   return remember {
     persistentListOf(
       selfLessonDecoration, // 自己的课程
